Modified layout.jsx for icons and also added a nav bar to be rendered above all the app's children:
import { config } from "@fortawesome/fontawesome-svg-core";
import "@fortawesome/fontawesome-svg-core/styles.css";

config.autoAddCss = false;

This code is utilizing Font Awesome, a popular icon library for websites and applications. Let's break down what each line does:

import { config } from "@fortawesome/fontawesome-svg-core";: This line imports the config object from the @fortawesome/fontawesome-svg-core package. This object is used to configure various settings related to Font Awesome.

import "@fortawesome/fontawesome-svg-core/styles.css";: This line imports the CSS styles for Font Awesome icons. The styles defined in this CSS file are necessary for displaying the icons correctly on the webpage.

config.autoAddCss = false;: This line configures Font Awesome to not automatically add CSS to the page. By default, Font Awesome automatically adds CSS to style the icons. However, setting autoAddCss to false prevents this behavior. This is useful if you want more control over how the icons are styled or if you're using a custom CSS framework.

Overall, this code snippet is configuring Font Awesome to not automatically add CSS styles to the page, allowing for more manual control over styling.

Nav bar styling:

flex-grow and overflow-y are CSS properties used for layout and styling purposes. Let's break down what each of them does:

flex-grow: This property specifies how much a flex item will grow relative to the rest of the flex items in the flex container. It determines the ability of a flex item to grow if necessary. The value is a unitless number that represents the proportion of the available space that the flex item should take up. For example, if one flex item has a flex-grow value of 2 and another has a value of 1, the first item will grow twice as much as the second item.

overflow-y: This property specifies how the content of an element should behave when it overflows the bounds of its container along the vertical axis (top to bottom). The value auto allows the browser to automatically add a vertical scrollbar when the content overflows, but only if necessary. This means that a scrollbar will appear when the content exceeds the height of the container.

So, flex-grow is used in flexbox layouts to control how flex items grow to fill available space, and overflow-y: auto is used to enable vertical scrolling within a container when its content overflows vertically.