Navigation Links:
Inside the navigation bar, there are two navigation links represented by Font Awesome icons wrapped in Link components.
Each Link component points to a different route ("/" and "/TicketPage/new") in the Next.js application.
The icons (faHome and faTicket) are displayed using the FontAwesomeIcon component, and their respective classes are assigned for styling.
User Information:

Next to the navigation links, there's a <div> containing the user's email address (jake.lower17@gmail.com). This information is likely to be dynamically populated based on the logged-in user.
The email address is styled using the class text-default-text.