GET Function:
This function is an asynchronous function that handles HTTP GET requests.
It expects two parameters: request and { params }. The params object likely contains route parameters extracted from the URL.
It extracts the id parameter from the params object.
It attempts to find a ticket in the database with the specified _id (assuming MongoDB-like _id field) using the Ticket.findOne() method.
If the ticket is found, it returns a JSON response containing the ticket data with a status code of 200 (OK).