Model Creation:
const Ticket = mongoose.models.Ticket || mongoose.model("Ticket", ticketSchema);: This line creates a Mongoose model named "Ticket" based on the defined schema. If the model already exists (i.e., if it was previously created), it uses the existing model; otherwise, it creates a new model.
Models in Mongoose represent collections in the MongoDB database. They provide an interface for querying and manipulating data in the collection.