Progress Bar Container:
The div has the following classes and attributes:
className="w-full bg-gray-200 rounded-full h-2.5 dark:bg-gray-700": These classes define the appearance of the progress bar container.
w-full: Makes the container occupy the full width of its parent element.
bg-gray-200: Sets the background color of the container to a light gray color.
rounded-full: Rounds the corners of the container to make it appear as a rounded rectangle.
h-2.5: Sets the height of the container to 2.5 units.
dark:bg-gray-700: This class applies a different background color for dark mode, changing the background to a darker shade of gray.