Progress Bar:
Inside the progress bar container, there's another div element representing the progress bar itself.
This div has the following attributes:
className="bg-blue-600 h-2.5 rounded-full": These classes define the appearance of the progress bar.
bg-blue-600: Sets the background color of the progress bar to a shade of blue.
h-2.5: Sets the height of the progress bar to 2.5 units.
rounded-full: Rounds the corners of the progress bar to make it appear as a rounded rectangle.
style={{ width: ${progress}% }}: This inline style dynamically sets the width of the progress bar based on the progress prop. The width is calculated as a percentage of the total width of the progress bar container.