getColor Function:
Inside the component, there's a helper function named getColor, which determines the background color of the status label based on the provided status prop.
It uses a switch statement to check the value of the status prop in a case-insensitive manner.
Depending on the value of status, it assigns a specific background color class:
If status is "done", it assigns the class "bg-green-200".
If status is "started", it assigns the class "bg-yellow-200".
If status is "not started", it assigns the class "bg-red-200".
If status doesn't match any of the above cases, it assigns the default class "bg-slate-700".
The function returns the determined color class.