getTicketById Function:
This is an asynchronous function named getTicketById.
It takes an id parameter, which represents the ID of the ticket to fetch.
Inside the function:
It makes a fetch request to the specified URL (http://localhost:3000/api/Tickets/${id}) to fetch the ticket data.
It sets cache: "no-store" in the fetch options to ensure that the response is not cached by the browser.
If the fetch operation is successful (res.ok), it returns the JSON response containing the ticket data.
If an error occurs during the fetch operation, it catches the error, logs it to the console, and returns undefined.