Form Submission Handling:
handleSubmit function handles form submission:
If in edit mode (EDITMODE is true), it sends a PUT request to update the ticket data using the ticket ID.
If in create mode (EDITMODE is false), it sends a POST request to create a new ticket.
After submission, it navigates the user back to the homepage using router.push("/") and refreshes the page with router.refresh().