Import Statements:
The code imports necessary modules and functions:
faX from @fortawesome/free-solid-svg-icons: This is the icon representing a delete action (usually an 'X' or close icon).
FontAwesomeIcon from @fortawesome/react-fontawesome: This is the component used to render FontAwesome icons.
useRouter from next/navigation: This hook is used for navigation and refreshing the page after deletion.